BE/requirements.txt
bcrypt==4.3.0
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
httptools==0.6.4
Jinja2==3.1.6
pip==25.0.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
SQLAlchemy-Utils==0.42.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
moviepy
opencv-python
